Style: added disabled button functionality. Feat: added conditionals to reserve a motorbike
In order to have a disable functionality which will make grayer the button, i need to change the class of the button to 'btn'.
Since that changed the shape of the button, i changed the other buttons to the same class.

I added some conditionals, functions and a conditional to prepared the code to control the flow when a client can or cannot reserve a
 motorbike.