Style: Change to a nav-bar the filters
I created a nav-bar with title _Filters_.
The content is a div with the previous part of the code of the div of filters.
Just changed colours and margins.

Margin left is hardcoded item by item since i had problems making the row of checkboxes centered.

Code can be impove.